moved Dashboard component inside layout folder
